# Performance improvements rasterization