--- a/src/Rasterizer.cpp
+++ b/src/Rasterizer.cpp
@@ -78,48 +78,67 @@ void Rasterizer::drawTriangle(std::array<Vector4f, 3> vertices, Shader* shader, 
     int maxy = std::min(std::max({y0, y1, y2}), frameBuffer->height - 1);
 
     float area = 1.f / Rasterizer::edgeCheck(x0, y0, x1, y1, x2, y2);
-    for (int i = minx; i <= maxx; i++) 
-    {
-        for (int j = miny; j <= maxy; j++) 
-        {
 
+    int edgeA = Rasterizer::edgeCheck(x1, y1, x2, y2, minx, miny);
+    int edgeB = Rasterizer::edgeCheck(x2, y2, x0, y0, minx, miny);
+    int edgeC = Rasterizer::edgeCheck(x0, y0, x1, y1, minx, miny);
+    int initialEdgeA, initialEdgeB, initialEdgeC;
+    int dxA = x2 - x1;
+    int dxB = x0 - x2;
+    int dxC = x1 - x0;
+    int dyA = y2 - y1;
+    int dyB = y0 - y2;
+    int dyC = y1 - y0;
+
+    for (int y = miny; y <= maxy; y++)
+    {
+        initialEdgeA = edgeA;
+        initialEdgeB = edgeB;
+        initialEdgeC = edgeC;
+
+        for (int x = minx; x <= maxx; x++)
+        {
             // v0 maps to edge v1v2
             // v1 maps to edge v2v0
             // v2 maps to edge v0v1
-            std::array<float, 3> weights{ Rasterizer::edgeCheck(x1, y1, x2, y2, i, j),
-                                          Rasterizer::edgeCheck(x2, y2, x0, y0, i, j),
-                                          Rasterizer::edgeCheck(x0, y0, x1, y1, i, j)};
+            std::array<float, 3> weights{(float)initialEdgeA, (float)initialEdgeB, (float)initialEdgeC};
 
             // skip if we are not in triangle
             // CCW order= negative is inside / positive is outside (we are using this)
             // CW order = negative is outside/ positive is inside
-            if (weights[0] > 0 || weights[1] > 0 || weights[2] > 0)
+            if (weights[0] <= 0 && weights[1] <= 0 && weights[2] <= 0)
             {
-                continue;
+
+                // normalize weights
+                // w0 + w1 + w2 = 1
+                weights[0] *= area;
+                weights[1] *= area;
+                weights[2] *= area;
+
+                float depth = 1.f/(weights[0] * z0 + weights[1] * z1 + weights[2] * z2);
+
+                if (depth <= depthBuffer->get(x, y))
+                {
+
+                    // perspective correct texture mapping
+                    weights[0] = weights[0] * z0 * depth;
+                    weights[1] = weights[1] * z1 * depth;
+                    weights[2] = weights[2] * z2 * depth;
+
+                    depthBuffer->set(x, y, depth);
+                    const std::array<uint8_t, 3> color = shader->processFragment(weights);
+                    frameBuffer->set(x, y, SDL_MapRGB(Rasterizer::PIXEL_FORMAT, color[0], color[1], color[2]));
+                }
             }
 
-            // normalize weights
-            // w0 + w1 + w2 = 1
-            weights[0] *= area;
-            weights[1] *= area;
-            weights[2] *= area;
-
-            float depth = 1.f/(weights[0] * z0 + weights[1] * z1 + weights[2] * z2);
-
-            if (depth > depthBuffer->get(i, j))
-            {
-                continue;
-            }
-
-            // perspective correct texture mapping
-            weights[0] = weights[0] * z0 * depth;
-            weights[1] = weights[1] * z1 * depth;
-            weights[2] = weights[2] * z2 * depth;
-
-            depthBuffer->set(i, j, depth);
-            const std::array<uint8_t, 3> color = shader->processFragment(weights);
-            frameBuffer->set(i, j, SDL_MapRGB(Rasterizer::PIXEL_FORMAT, color[0], color[1], color[2]));
+            initialEdgeA += dyA;
+            initialEdgeB += dyB;
+            initialEdgeC += dyC;
         }
+
+        edgeA -= dxA;
+        edgeB -= dxB;
+        edgeC -= dxC;
     }
 }