--- a/src/TextureMapper.cpp
+++ b/src/TextureMapper.cpp
@@ -2,10 +2,10 @@
 
 uint32_t TextureMapper::sample(const TextureBuffer* textureBuffer, 
                                const std::array<Vector4f, 3>& textureVertices,
-                               const std::array<float, 3>& weights)
+                               float w0, float w1, float w2)
 {
-    float u = textureVertices[0].x * weights[0] + textureVertices[1].x * weights[1] + textureVertices[2].x * weights[2];
-    float v = textureVertices[0].y * weights[0] + textureVertices[1].y * weights[1] + textureVertices[2].y * weights[2];
+    float u = textureVertices[0].x * w0 + textureVertices[1].x * w1 + textureVertices[2].x * w2;
+    float v = textureVertices[0].y * w0 + textureVertices[1].y * w1 + textureVertices[2].y * w2;
     u = std::min(u, 1.f);
     v = std::min(v, 1.f);
     int width = u * textureBuffer->width - 1;