This PR updates the code with a few small things but the performance improved a lot. Rendering the initial scene with two cubes was taking `67ms`.

Change | Before | After
-|-|-|
Moved pixel discarding logic before the loop | 67ms | 64ms
Updated the rasterization loop to compute edge checks iteratively | 64ms | 60ms
Updated the fragment shader to return `uint32_t` instead of an array of `uint8_t` | 60ms | 50ms
Started passing barycentric coords as separate variables instead of an array | 50ms | 18ms
Started passing diffuse texture vertices as separate variables instead of an array | 18ms | 9ms

Currently the initial scene is being rendered in `9ms`. Getting close to the cubes slows down the rendering but that is expected as the rasterization area gets larger. Will deal with this in the future.